Add 1.15.0 news post [WIP]
Note: This update is not published yet, and the changelog isn't final. It's provided in advance to have a better coverage across languages and less burden on translators when the update comes out.

We'll probably include a flavor text at the beginning explaining why we consider this to be a major update, haven't written one yet.

Open beta announcement will be a separate post that we'll commit at some point.

Feel free to share the news in advance if you like, this isn't confidential. :P